Correct `aria-controls` to refernce correct id
Issue was caught by aXe-core testing: https://dequeuniversity.com/rules/axe/4.4/aria-valid-attr-value?application=axe-puppeteer

The issue happened because `tag_group_name` was added to the beginning of each id / href, but not to aria-controls.
The `tab_id` was never used individually, so I just updated it to include both `tab_group_name` and `tab_id`,
and then replaced its used in the tabs and tab panels.