Make collapse/display templates work in dark mode
In dark mode, the display templates still have a light background, and stand
out far too much. This adds CSS to change the `bg-light` class for the collapse
and display templates to the dark-mode colors when `data-bs-theme=dark`.

See https://github.com/jhpyle/docassemble/pull/718 for a discussion on a
suggested fix for a similar issue, which resulted in the suggested workaround
here.